+++ src/components/molecules/JoinUs/index.tsx
@@ -1,4 +1,5 @@
 import { useTranslation } from "react-i18next";
+import { DISCORD_URL } from "src/consts";
 import DiscordIcon from "src/icons/DiscordIcon";
 import "./styles.scss";
 
@@ -12,9 +13,11 @@ const JoinUs = () => {
           {t("home.join.initText")}
           <span> {t("home.join.endText")}</span>
         </div>
-        <button className="home-join-button" data-testid="join-discord-button">
-          <DiscordIcon /> {t("home.join.discord")}
-        </button>
+        <a href={DISCORD_URL} target="_blank" rel="noreferrer">
+          <button className="home-join-button" data-testid="join-discord-button">
+            <DiscordIcon /> {t("home.join.discord")}
+          </button>
+        </a>
       </div>
     </section>
   );
