+++ src/pages/Txs/Information/styles.scss
@@ -17,13 +17,11 @@
     color: var(--color-primary-50);
   }
 
+  .tabs {
+    overflow: auto;
+  }
+
   .table {
-    display: none;
-
-    @include desktop {
-      display: table;
-    }
-
     & a {
       color: var(--color-information-100);
       text-decoration: none;
